Add files via upload
This potential change is a modification to the inputs used in the TAX_CALCULATOR function. Before this, the function would support any input, but would not verify that the input could be used properly resulting in a large number of errors. With this change, the inputs are first 'verified' before being used in the function. In addition, this contains 'checks' that ensure the user is inputting correct values into the function.