Applied additional changes to Function-Modifiers
I've updated each function to include the 'check' that  was implemented
at the top. In addition, this change modifies the COIN_FLIP function.
Previously the COIN_FLIP function would only display the result for each
flip. With this new change the total amounts of heads and tails are
counted and displayed at the end. In addition the time between flips was
decreased by 1 second as 2 seconds took much too long.